+++ src/features/XIT/CHAT.vue
@@ -32,28 +32,41 @@ watchEffect(() => {
   <div v-if="!parameter">Error! Not Enough Parameters!</div>
   <LoadingSpinner v-else-if="!isLoaded" />
   <div v-else :style="{ height: '100%', flexGrow: 1, paddingTop: '4px' }">
-    <div class="title">{{ parameter }} Global Site Owners</div>
-    <div v-for="(message, i) in messages" :key="i" :class="$style.line">
-      <div>
-        <div :class="$style.date">{{ ddmm(message.MessageTimestamp) }}</div>
-        <div :class="$style.date" :style="{ color: '#999999' }">
+    <div :class="$style.title">{{ parameter }} Global Site Owners</div>
+    <div
+      v-for="(message, i) in messages"
+      :key="i"
+      :class="[C.Message.message, C.type.typeRegular, C.fonts.fontRegular]">
+      <div :class="C.Message.timestamp">
+        <div :class="C.Message.date">{{ ddmm(message.MessageTimestamp) }}</div>
+        <div :class="C.Message.time" :style="{ color: '#999999' }">
           {{ hhmm(message.MessageTimestamp) }}
         </div>
       </div>
       <template v-if="message.MessageType === 'CHAT'">
-        <div :class="$style.name">{{ message.UserName }}</div>
-        <div :class="$style.message">{{ message.MessageText }}</div>
+        <div :class="C.Message.name">
+          <div :class="[C.Sender.container, C.type.typeRegular, C.fonts.fontRegular]">
+            <div :class="[C.Sender.name, $style.name]">
+              {{ message.UserName }}
+            </div>
+          </div>
+        </div>
+        <div :class="C.Message.controlsAndText">
+          <div :class="C.Message.text">
+            {{ message.MessageText }}
+          </div>
+        </div>
       </template>
       <template v-if="message.MessageType === 'JOINED'">
-        <div :class="$style.name">{{ message.UserName }}</div>
-        <div :class="$style.message" :style="{ fontStyle: 'italic' }">
-          {{ message.UserName }} joined.
+        <div :class="C.Message.name" />
+        <div :class="C.Message.controlsAndText">
+          <div :class="[C.Message.text, C.Message.system]">{{ message.UserName }} joined.</div>
         </div>
       </template>
       <template v-if="message.MessageType === 'LEFT'">
-        <div :class="$style.name" />
-        <div :class="$style.message" :style="{ fontStyle: 'italic' }">
-          {{ message.UserName }} left.
+        <div :class="C.Message.name" />
+        <div :class="C.Message.controlsAndText">
+          <div :class="[C.Message.text, C.Message.system]">{{ message.UserName }} left.</div>
         </div>
       </template>
     </div>
@@ -61,39 +74,14 @@ watchEffect(() => {
 </template>
 
 <style module>
-.line {
-  width: 100%;
-  display: grid;
-  grid-template-columns: minmax(8em, 1fr) minmax(8em, 4fr) minmax(8em, 15fr);
-  grid-column-gap: 1px;
-  font-size: 11px;
-  line-height: 1.1;
-}
-
-.date {
-  color: #444444;
-  display: inline-block;
-  padding: 2px 5px;
-  padding-right: 0px;
-  text-align: left;
-  white-space: nowrap;
-  overflow: hidden;
+.title {
+  font-weight: bold;
+  display: block;
+  font-size: 16px;
+  padding-left: 5px;
 }
 
 .name {
-  display: inline-block;
-  text-align: right;
-  color: #999999;
-  text-overflow: ellipsis;
-  padding: 2px 5px;
-  border-right: 1px solid #999999;
-}
-
-.message {
-  display: inline-block;
-  text-align: left;
-  color: #bbbbbb;
-  word-break: break-word;
-  padding: 2px 5px;
+  cursor: text;
 }
 </style>